move all scoring algorithms to scores file
