+++ client/modules/espells/src/suggest/ngram.ts
@@ -3,8 +3,14 @@ import type { PrefixMap, SuffixMap } from "../aff"
 import type { Prefix, Suffix } from "../aff/affix"
 import { CONSTANTS as C } from "../constants"
 import type { Word } from "../dic/word"
-import { commonCharacters, lcslen, leftCommonSubstring, lowercase, ngram } from "../util"
-import { rootScore, ScoresList } from "./scores"
+import { lowercase } from "../util"
+import {
+  preciseAffixScore,
+  rootScore,
+  roughAffixScore,
+  ScoresList,
+  scoreThreshold
+} from "./scores"
 
 export function* ngramSuggest(
   misspelling: string,
@@ -32,7 +38,7 @@ export function* ngramSuggest(
     roots.add(score, word)
   }
 
-  const threshold = detectThreshold(misspelling)
+  const threshold = scoreThreshold(misspelling)
 
   const guesses = new ScoresList<[string, string]>(C.NGRAM_MAX_GUESSES)
 
@@ -62,68 +68,6 @@ export function* ngramSuggest(
   )
 
   yield* filterGuesses(guesses2, known, onlyMaxDiff)
-}
-
-function roughAffixScore(word1: string, word2: string) {
-  return (
-    ngram(word1.length, word1, word2, false, true) + leftCommonSubstring(word1, word2)
-  )
-}
-
-function preciseAffixScore(
-  word1: string,
-  word2: string,
-  diffFactor: number,
-  base: number,
-  hasPhonetic: boolean
-) {
-  const lcs = lcslen(word1, word2)
-
-  if (word1.length === word2.length && word1.length === lcs) {
-    return base + 2000
-  }
-
-  let result: number
-
-  result = 2 * lcs - Math.abs(word1.length - word2.length)
-
-  result += leftCommonSubstring(word1, word2)
-
-  if (commonCharacters(word1, lowercase(word2))) result++
-
-  result += ngram(4, word1, word2, false, true)
-
-  const bigrams = ngram(2, word1, word2, true, true) + ngram(2, word2, word1, true, true)
-
-  result += bigrams
-
-  let questionableLimit: number
-  if (hasPhonetic) {
-    questionableLimit = word1.length * diffFactor
-  } else {
-    questionableLimit = (word1.length + word2.length) * diffFactor
-  }
-
-  if (bigrams < questionableLimit) result -= 1000
-
-  return result
-}
-
-function detectThreshold(word: string) {
-  let threshold = 0
-
-  for (let startPos = 1; startPos < 4; startPos++) {
-    const mangled: string[] = []
-    for (let pos = startPos; pos < word.length; pos += 4) {
-      mangled[pos] = "*"
-    }
-
-    const mangledWord = mangled.join("")
-
-    threshold += ngram(word.length, word, mangledWord, false, true)
-  }
-
-  return Math.floor(threshold / (3 - 1))
 }
 
 function formsFor(
